BUILD: debug: fix build issue on clang with previous commit
Since the thread_dump_state type changed to uint, the old value in the
CAS needs to be the same as well.

+++ src/debug.c
@@ -1311,7 +1311,7 @@ struct buffer *thread_dump_buffer = NULL;
 /* initiates a thread dump */
 void ha_thread_dump_all_to_trash()
 {
-	unsigned long old;
+	unsigned int old;
 
 	while (1) {
 		old = 0;
